begun refactoring of outs methods into their own module currently writting test coverage over the functions.outs module have not updated texas_holdem.py to point to new location have not bugun refactoring of outs_tb methods to new module
