+++ src/pypoker/engine/hand_solver/texas_holdem.py
@@ -21,6 +21,7 @@ from pypoker.engine.hand_solver.constants import (
     HAND_TYPE_TRIPS, HAND_TYPE_TWO_PAIR, HAND_TYPE_PAIR, HAND_TYPE_HIGH_CARD, GAME_TYPE_TEXAS_HOLDEM,
 )
 from pypoker.engine.hand_solver.functions import hand_test, rank_hand_type, describe_hand
+from pypoker.engine.hand_solver.utils import get_all_combinations
 
 
 class TexasHoldemHandSolver(BaseHandSolver):
@@ -129,7 +130,7 @@ class TexasHoldemHandSolver(BaseHandSolver):
             if len(hole_cards) + len(board_cards) >= 5
             else len(hole_cards) + len(board_cards)
         )
-        all_hands = self.get_all_combinations(hole_cards, board_cards, hand_size)
+        all_hands = get_all_combinations(hole_cards, board_cards, hand_size)
         for hand_ranking in self._hand_rankings:
             matched_hands = [hand for hand in all_hands if hand_test(GAME_TYPE_TEXAS_HOLDEM, hand_ranking[HAND_TITLE],
                                                                      hand=hand)]
@@ -323,417 +324,6 @@ class TexasHoldemHandSolver(BaseHandSolver):
             player: win_count / draw_combo_num for player, win_count in wins.items()
         }
         return odds
-
-
-    #########################
-    #  OUTS FINDER METHODS  #
-    #########################
-
-    def _outs_straight_flush(self, hole_cards, board_cards, available_cards):
-        """
-        Private method to find all of the possible draw scenarios to get the player this specific draw type.
-        :param hole_cards: List of card objects representing the players hole cards
-        :param board_cards: List of card objects representing the cards that have been dealt to the board.
-        :param available_cards: List of card objects representing the cards that can be drawn.
-        """
-
-        draw_scenarios = []
-        draws = 5 - len(board_cards)
-        drawn_cards = hole_cards + board_cards
-        suit_drawn = {
-            SUIT_HEARTS: sum([1 for card in drawn_cards if card.suit == SUIT_HEARTS]),
-            SUIT_CLUBS: sum([1 for card in drawn_cards if card.suit == SUIT_CLUBS]),
-            SUIT_DIAMONDS: sum(
-                [1 for card in drawn_cards if card.suit == SUIT_DIAMONDS]
-            ),
-            SUIT_SPADES: sum([1 for card in drawn_cards if card.suit == SUIT_SPADES]),
-        }
-        suit_remaining = {
-            SUIT_HEARTS: sum(
-                [1 for card in available_cards if card.suit == SUIT_HEARTS]
-            ),
-            SUIT_CLUBS: sum([1 for card in available_cards if card.suit == SUIT_CLUBS]),
-            SUIT_DIAMONDS: sum(
-                [1 for card in available_cards if card.suit == SUIT_DIAMONDS]
-            ),
-            SUIT_SPADES: sum(
-                [1 for card in available_cards if card.suit == SUIT_SPADES]
-            ),
-        }
-
-        suit_eligable = [
-            suit
-            for suit, count in suit_drawn.items()
-            if count + draws >= 5 and suit_remaining[suit] >= (5 - count)
-        ]
-
-        for suit in suit_eligable:
-            drawn_suit_cards = [card for card in drawn_cards if card.suit == suit]
-            available_suit_cards = [
-                card for card in available_cards if card.suit == suit
-            ]
-
-            drawn_suit_cards = sorted(
-                drawn_suit_cards, key=lambda card: card.value, reverse=True
-            )
-            available_suit_cards = sorted(
-                available_suit_cards, key=lambda card: card.value, reverse=True
-            )
-
-            for draws_used in range(1, draws + 1):
-                for hypothetical_draw_cards in self.get_all_combinations(
-                    available_suit_cards, [], draws_used
-                ):
-                    usable_cards = drawn_suit_cards + hypothetical_draw_cards
-                    usable_card_vals = [card.value for card in usable_cards]
-                    usable_card_vals.sort()
-
-                    gaps = [
-                        [s, e]
-                        for s, e in zip(usable_card_vals, usable_card_vals[1:])
-                        if s + 1 < e
-                    ]
-                    edges = iter(
-                        usable_card_vals[:1] + sum(gaps, []) + usable_card_vals[-1:]
-                    )
-                    runs = list(zip(edges, edges))
-
-                    qualifying_straights = [
-                        run
-                        for run in runs
-                        if run[1] - run[0] >= 4
-                        and all(
-                            (run[1] - 4) <= card.value <= run[1]
-                            for card in hypothetical_draw_cards
-                        )
-                    ]
-
-                    for straight_vals in qualifying_straights:
-                        suits = []
-                        values = []
-                        for card in hypothetical_draw_cards:
-                            suits.append(card.suit)
-                            values.append(card.value)
-                        draw_scenarios.append(
-                            {
-                                OUT_STRING: self.build_out_string(suits, values, draws),
-                                TIEBREAKER: straight_vals[1],
-                            }
-                        )
-
-        return draw_scenarios
-
-    def _outs_quads(self, hole_cards, board_cards, available_cards):
-        """
-        Private method to find all of the possible draw scenarios to get the player this specific draw type.
-        :param hole_cards: List of card objects representing the players hole cards
-        :param board_cards: List of card objects representing the cards that have been dealt to the board.
-        :param available_cards: List of card objects representing the cards that can be drawn.
-        """
-
-        draw_scenarios = []
-        draws = 5 - len(board_cards)
-        drawn_cards = hole_cards + board_cards
-
-        drawn_value_counter = Counter([card.value for card in drawn_cards])
-        available_values_counter = Counter([card.value for card in available_cards])
-        value_eligable = [
-            value
-            for value, count in drawn_value_counter.items()
-            if count + draws >= 4 and available_values_counter[value] + value >= 4
-        ]
-
-        for value in value_eligable:
-            cards_needed = 4 - drawn_value_counter[value]
-            suits = ["ANY" for _ in range(cards_needed)]
-            values = [value for _ in range(cards_needed)]
-
-            draw_scenarios.append(
-                {
-                    OUT_STRING: self.build_out_string(suits, values, draws),
-                    TIEBREAKER: value,
-                }
-            )
-
-        return draw_scenarios
-
-    def _outs_full_house(self, hole_cards, board_cards, available_cards):
-        """
-        Private method to find all of the possible draw scenarios to get the player this specific draw type.
-        :param hole_cards: List of card objects representing the players hole cards
-        :param board_cards: List of card objects representing the cards that have been dealt to the board.
-        :param available_cards: List of card objects representing the cards that can be drawn.
-        """
-
-        draw_scenarios = []
-        draws = 5 - len(board_cards)
-        drawn_cards = hole_cards + board_cards
-
-        drawn_value_counter = Counter([card.value for card in drawn_cards])
-        available_values_counter = Counter([card.value for card in available_cards])
-
-        trips_eligable = [
-            (value, drawn_value_counter[value])
-            for value in CARD_VALUES
-            if drawn_value_counter[value] + draws >= 3
-            and available_values_counter[value] + drawn_value_counter[value] >= 3
-        ]
-
-        pair_eligable = [
-            (value, drawn_value_counter[value])
-            for value in CARD_VALUES
-            if drawn_value_counter[value] + draws >= 2
-            and available_values_counter[value] + drawn_value_counter[value] >= 2
-        ]
-
-        for trip_info, pair_info in product(trips_eligable, pair_eligable):
-            # check we havent selected the same value for both trips and pairs
-            if trip_info[0] == pair_info[0]:
-                continue
-
-            trip_draws_required = max([3 - trip_info[1], 0])
-            pair_draws_required = max([2 - pair_info[1], 0])
-
-            if trip_draws_required + pair_draws_required > draws:
-                continue
-
-            suits = ["ANY" for _ in range(trip_draws_required + pair_draws_required)]
-            values = [trip_info[0] for _ in range(trip_draws_required)]
-            values.extend([pair_info[0] for _ in range(pair_draws_required)])
-
-            draw_scenarios.append(
-                {
-                    OUT_STRING: self.build_out_string(suits, values, draws),
-                    TIEBREAKER: (trip_info[0], pair_info[0]),
-                }
-            )
-        return draw_scenarios
-
-    def _outs_flush(self, hole_cards, board_cards, available_cards):
-        """
-        Private method to find all of the possible draw scenarios to get the player this specific draw type.
-        :param hole_cards: List of card objects representing the players hole cards
-        :param board_cards: List of card objects representing the cards that have been dealt to the board.
-        :param available_cards: List of card objects representing the cards that can be drawn.
-        """
-
-        draw_scenarios = []
-        draws = 5 - len(board_cards)
-        drawn_cards = hole_cards + board_cards
-        suit_drawn = {
-            SUIT_HEARTS: sum([1 for card in drawn_cards if card.suit == SUIT_HEARTS]),
-            SUIT_CLUBS: sum([1 for card in drawn_cards if card.suit == SUIT_CLUBS]),
-            SUIT_DIAMONDS: sum(
-                [1 for card in drawn_cards if card.suit == SUIT_DIAMONDS]
-            ),
-            SUIT_SPADES: sum([1 for card in drawn_cards if card.suit == SUIT_SPADES]),
-        }
-        suit_remaining = {
-            SUIT_HEARTS: sum(
-                [1 for card in available_cards if card.suit == SUIT_HEARTS]
-            ),
-            SUIT_CLUBS: sum([1 for card in available_cards if card.suit == SUIT_CLUBS]),
-            SUIT_DIAMONDS: sum(
-                [1 for card in available_cards if card.suit == SUIT_DIAMONDS]
-            ),
-            SUIT_SPADES: sum(
-                [1 for card in available_cards if card.suit == SUIT_SPADES]
-            ),
-        }
-
-        suit_eligable = [
-            suit
-            for suit, count in suit_drawn.items()
-            if count + draws >= 5 and suit_remaining[suit] >= (5 - count)
-        ]
-
-        for suit in suit_eligable:
-            draws_required = max([5 - suit_drawn[suit], 0])
-            suits = [suit for _ in range(draws_required)]
-            values = ["ANY" for _ in range(draws_required)]
-
-            draw_scenarios.append(
-                {
-                    OUT_STRING: self.build_out_string(suits, values, draws),
-                    TIEBREAKER: suit,
-                }
-            )
-
-        return draw_scenarios
-
-    def _outs_straight(self, hole_cards, board_cards, available_cards):
-        """
-        Private method to find all of the possible draw scenarios to get the player this specific draw type.
-        :param hole_cards: List of card objects representing the players hole cards
-        :param board_cards: List of card objects representing the cards that have been dealt to the board.
-        :param available_cards: List of card objects representing the cards that can be drawn.
-        """
-
-        draw_scenarios = []
-        draws = 5 - len(board_cards)
-        drawn_cards = hole_cards + board_cards
-
-        drawn_values = sorted(
-            list(set([card.value for card in drawn_cards])), reverse=True
-        )
-        available_values = sorted(
-            list(set([card.value for card in available_cards])), reverse=True
-        )
-        not_drawn_values = [
-            value for value in available_values if value not in drawn_values
-        ]
-
-        for draws_used in range(1, draws + 1):
-            for possible_draw in combinations(not_drawn_values, draws_used):
-                combined_values = sorted(drawn_values + list(possible_draw))
-
-                gaps = [
-                    [s, e]
-                    for s, e in zip(combined_values, combined_values[1:])
-                    if s + 1 < e
-                ]
-                edges = iter(combined_values[:1] + sum(gaps, []) + combined_values[-1:])
-                runs = list(zip(edges, edges))
-
-                qualifying_straights = [
-                    run
-                    for run in runs
-                    if run[1] - run[0] >= 4
-                    and all((run[1] - 4) <= value <= run[1] for value in possible_draw)
-                ]
-
-                for straight_vals in qualifying_straights:
-                    suits = ["ANY" for _ in possible_draw]
-                    values = [value for value in possible_draw]
-                    draw_scenarios.append(
-                        {
-                            OUT_STRING: self.build_out_string(suits, values, draws),
-                            TIEBREAKER: straight_vals[1],
-                        }
-                    )
-        return draw_scenarios
-
-    def _outs_trips(self, hole_cards, board_cards, available_cards):
-        """
-        Private method to find all of the possible draw scenarios to get the player this specific draw type.
-        :param hole_cards: List of card objects representing the players hole cards
-        :param board_cards: List of card objects representing the cards that have been dealt to the board.
-        :param available_cards: List of card objects representing the cards that can be drawn.
-        """
-
-        draw_scenarios = []
-        draws = 5 - len(board_cards)
-        drawn_cards = hole_cards + board_cards
-
-        drawn_value_counter = Counter([card.value for card in drawn_cards])
-        available_values_counter = Counter([card.value for card in available_cards])
-
-        trips_eligable = [
-            (value, drawn_value_counter[value])
-            for value in CARD_VALUES
-            if drawn_value_counter[value] + draws >= 3
-            and available_values_counter[value] + drawn_value_counter[value] >= 3
-        ]
-
-        for value, count in trips_eligable:
-            draws_required = max([3 - count, 0])
-            suits = ["ANY" for _ in range(draws_required)]
-            values = [value for _ in range(draws_required)]
-
-            draw_scenarios.append(
-                {
-                    OUT_STRING: self.build_out_string(suits, values, draws),
-                    TIEBREAKER: value,
-                }
-            )
-
-        return draw_scenarios
-
-    def _outs_two_pair(self, hole_cards, board_cards, available_cards):
-        """
-        Private method to find all of the possible draw scenarios to get the player this specific draw type.
-        :param hole_cards: List of card objects representing the players hole cards
-        :param board_cards: List of card objects representing the cards that have been dealt to the board.
-        :param available_cards: List of card objects representing the cards that can be drawn.
-        """
-
-        draw_scenarios = []
-        draws = 5 - len(board_cards)
-        drawn_cards = hole_cards + board_cards
-
-        drawn_value_counter = Counter([card.value for card in drawn_cards])
-        available_values_counter = Counter([card.value for card in available_cards])
-
-        pair_eligable = [
-            (value, drawn_value_counter[value])
-            for value in CARD_VALUES
-            if drawn_value_counter[value] + draws >= 2
-            and available_values_counter[value] + drawn_value_counter[value] >= 2
-        ]
-
-        for pair_a, pair_b in combinations(pair_eligable, 2):
-            high_pair = pair_a if pair_a[0] > pair_b[0] else pair_b
-            low_pair = pair_a if pair_a[0] < pair_b[0] else pair_b
-
-            high_pair_draws_req = max([2 - high_pair[1], 0])
-            low_pair_draws_req = max([2 - low_pair[1], 0])
-            if high_pair_draws_req + low_pair_draws_req <= draws:
-                suits = ["ANY" for _ in range(high_pair_draws_req + low_pair_draws_req)]
-                values = [high_pair[0] for _ in range(high_pair_draws_req)]
-                values.extend([low_pair[0] for _ in range(low_pair_draws_req)])
-                draw_scenarios.append(
-                    {
-                        OUT_STRING: self.build_out_string(suits, values, draws),
-                        TIEBREAKER: (high_pair[0], low_pair[0]),
-                    }
-                )
-
-        return draw_scenarios
-
-    def _outs_pair(self, hole_cards, board_cards, available_cards):
-        """
-        Private method to find all of the possible draw scenarios to get the player this specific draw type.
-        :param hole_cards: List of card objects representing the players hole cards
-        :param board_cards: List of card objects representing the cards that have been dealt to the board.
-        :param available_cards: List of card objects representing the cards that can be drawn.
-        """
-
-        draw_scenarios = []
-        draws = 5 - len(board_cards)
-        drawn_cards = hole_cards + board_cards
-
-        drawn_value_counter = Counter([card.value for card in drawn_cards])
-        available_values_counter = Counter([card.value for card in available_cards])
-
-        pair_eligable = [
-            (value, drawn_value_counter[value])
-            for value in CARD_VALUES
-            if drawn_value_counter[value] + draws >= 2
-            and available_values_counter[value] + drawn_value_counter[value] >= 2
-        ]
-
-        for pair in pair_eligable:
-            draws_req = max([2 - pair[1], 0])
-            suits = ["ANY" for _ in range(draws_req)]
-            values = [pair[0] for _ in range(draws_req)]
-            draw_scenarios.append(
-                {
-                    OUT_STRING: self.build_out_string(suits, values, draws),
-                    TIEBREAKER: pair[0],
-                }
-            )
-
-        return draw_scenarios
-
-    def _outs_high_card(self, hole_cards, board_cards, available_cards):
-        """
-        Private method to find all of the possible draw scenarios to get the player this specific draw type.
-        :param hole_cards: List of card objects representing the players hole cards
-        :param board_cards: List of card objects representing the cards that have been dealt to the board.
-        :param available_cards: List of card objects representing the cards that can be drawn.
-        """
-
-        return []
 
     #############################
     #  OUTS TIEBREAKER METHODS  #
