+++ tests/unit/engine/hand_solver/functions/test_outs.py
@@ -1,0 +1,35 @@
+from pytest import mark
+
+from pypoker.engine.hand_solver.functions.outs import build_out_string, claim_out_strings
+
+
+@mark.parametrize("suits, values, draws, expected", [
+    (["Spades"], [2], 5, "S2-**-**-**-**"),
+    (["Spades", "Hearts"], [2, 14], 5, "S2-HA-**-**-**"),
+    (["Spades", "Hearts", "Clubs"], [2, 14, 3], 5, "S2-HA-C3-**-**"),
+    (["Spades", "Hearts", "Clubs", "Diamonds"], [2, 14, 3, 13], 5, "S2-HA-C3-DK-**"),
+    (["Spades", "Hearts", "Clubs", "Diamonds", "Spades"], [2, 14, 3, 13, 4], 5, "S2-HA-C3-DK-S4"),
+    (["Spades"], [5], 4, "S5-**-**-**"),
+    (["Spades", "Hearts"], [5, 12], 4, "S5-HQ-**-**"),
+    (["Spades", "Hearts", "Clubs"], [5, 12, 6], 4, "S5-HQ-C6-**"),
+    (["Spades", "Hearts", "Clubs", "Diamonds"], [5, 12, 6, 11], 4, "S5-HQ-C6-DJ"),
+    (["Spades"], [7], 3, "S7-**-**"),
+    (["Spades", "Hearts"], [7, 10], 3, "S7-HT-**"),
+    (["Spades", "Hearts", "Clubs"], [7, 10, 8], 3, "S7-HT-C8"),
+    (["Spades"], [9], 2, "S9-**"),
+    (["Spades", "Hearts"], [9, 14], 2, "S9-HA"),
+    (["Spades"], [9], 1, "S9"),
+    (["Hearts", "Hearts", "Hearts"], ["ANY", "ANY", "ANY"], 5, "H*-H*-H*-**-**"),
+    (["ANY", "ANY", "ANY", "ANY"], [6, 6, 7, 7], 5, "*6-*6-*7-*7-**")
+])
+def test_when_build_out_string_then_correct_string_returned(suits, values, draws, expected):
+    actual = build_out_string(suits, values, draws)
+    assert actual == expected
+
+
+@mark.parametrize("utilised_outs, out_strings, drawable_cards, expected", [
+    ([], ["S*"], "drawable_cards_001", [])
+])
+def test_when_claim_out_strings_then_correct_values_claimed(utilised_outs, out_strings, drawable_cards, expected):
+    actual = claim_out_strings(utilised_outs, out_strings, drawable_cards)
+    assert actual == expected
