+++ tests/unit/engine/hand_solver/test_utils.py
@@ -2,7 +2,7 @@ from pytest import mark
 
 from fixtures.cards import get_hand, get_hand_sets
 from pypoker.engine.hand_solver.utils import hand_all_same_suit, hand_values_continuous, hand_highest_value_tuple, \
-    hand_is_ace_low_straight, order_hands_highest_card, hands_have_same_card_values
+    hand_is_ace_low_straight, order_hands_highest_card, hands_have_same_card_values, get_all_combinations
 
 
 @mark.parametrize("hand, expected", [
@@ -90,3 +90,19 @@ def test_when_hands_have_same_card_values_then_correct_response_returned(hand_a,
     actual = hands_have_same_card_values(hand_a, hand_b)
 
     assert actual == expected
+
+
+@mark.parametrize("hole_cards, board_cards, hand_size, expected_combos", [
+    ("hole_cards_001", "board_cards_001", 5, "combos_001"),
+    ("hole_cards_001", "board_cards_002", 5, "combos_002"),
+    ("hole_cards_001", "board_cards_003", 5, "combos_003")
+])
+def test_when_get_all_combinations_then_correct_combinations_returned(hole_cards, board_cards, hand_size,
+                                                                      expected_combos):
+    hole_cards = get_hand(hole_cards)
+    board_cards = get_hand(board_cards)
+    expected_combos = get_hand_sets(expected_combos)
+
+    combos = get_all_combinations(hole_cards, board_cards, hand_size)
+
+    assert combos == expected_combos
